The code to replace $wid/$pid died on $foo.
It failed to increment a variable after finding a $ but no pid or wid,
which led to a fun infinite loop. Also don't check if c[1] is w if we
already established it was p. The code should probably check the length
of the string instead of reading random memory, but how fun is that?