"Fix" releasing the mouse button over a window.
Previously if you click-drag-released the left button on the root window to a client window,
Openbox would ignore the release because we don't own the client window. Now check if we
are in a button press or not before ignoring it. The bug was we would initiate a move when
the cursor was moved over a titlebar even when the button was unpressed because the variable
didn't get cleared properly.

This fix is very ugly.