Directional switching
On 2015-07-07 13:17 -0700, Ian Zimmerman wrote:

> The 2 and 5 could/should be rc.xml parameters.  With defaults 1 and 1
> you'd have the current behavior.  Is there interest in me pursuing that
> path and providing a more complete patch?

Not much interest, but here it is anyway.  I want to be on record as
using this feature to its full extent, and if necessary maintaining it.

--
Please *no* private copies of mailing list or newsgroup messages.
Rule 420: All persons more than eight miles high to leave the court.