Fix an invalid shift and remove a redundant NULL check in openbox/event.c
In function event_handle_client, throughout the code the variable
client is assumed not to be NULL.
On line 1077, its 'frame' member is passed to a function, thus
assuming it to be non-NULL.
A few lines after, at 1084, a ternary operator asks whether
client is or isn't NULL.
One of those is wrong.
I opted for assuming client to not be NULL because that's what
everyone else in the function assumes.
On the same file, a (1 << 31) was performed, which is undefined,
since the literal 1 is signed. I simply changed it to (1U << 31),
because that's consistent with the types used.