Implements #12
Added a new XML node into ForEach action configuration - `none`.

Extracted application of `if` action into separate function which
returns result that indicates of a client matches the query.

`ForEach` tracks the execution of `do_run_if` and exectues actions
mentioned in `none` node if none of the `if` were able to find a match