Add "applyDefaultIcon" setting
Some applications don't provide their own icons. By default, Openbox
sets a generic icon for these applications. Other environments, such
as GNOME, instead supply an icon from /usr/share/pixmaps or the icon
theme. Openbox's behavior is appropriate when it is running on its
own, but when it being used as the window manager in a broader
desktop environment, the presence of the default icon prevents the
environment from noticing that it needs to provide an icon.

This change adds an `<applyDefaultIcon>` configuration option that
defaults to `yes` to preserve the standard Openbox behavior. By
setting this option to `no`, Openbox will cede responsibility for
setting default icons to the desktop environment.

Resolves issue #21